task2/requirements.txt
Flask
psycopg2
pgvector
python-dotenv
sentence-transformers
PyPDF2
openai
